% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{leerConfig}
\alias{leerConfig}
\title{Title : Leer Config}
\usage{
leerConfig(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Title : Leer Config
}
\examples{

}
\author{
Kirsten/Ernesto/Borja
}
